Experimenting with the simple spark job and its serialization
There appears to be something problematic about this Proc as it's coming through
the serialization. Other DataSet functions work fine, but there's something
about filter() which isn't quite working.

Since most of what we want to do is with maps and filters, I suppose figuring
this out will be important